+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+# 1.0.1 (2020-06-30)
+
+### Bugfixes
+
+- Fixed a bug that generates an error when an user has only one group associated with his account. That means that this one group is specified as a string and not as an Array.
+
 # 1.0.0 (2020-06-25)
 
 ### Features
 
-* Publishing core version of `middy-cognito-groups-authorizer`
+- Publishing core version of `middy-cognito-groups-authorizer`